% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{eval_log}
\alias{eval_log}
\alias{eval_log_read}
\alias{eval_log_write}
\alias{eval_log_dir}
\alias{eval_log_dir_set}
\title{Read and write evaluation logs}
\usage{
eval_log_read(file)

eval_log_write(x = eval_log_new(), dir = eval_log_dir())

eval_log_dir()

eval_log_dir_set(dir)
}
\description{
Read and write evaluation logs
}
\examples{
file <-
  system.file(
    "logs/2025-02-08T15-51-00-06-00_simple-arithmetic_o3cKtmsvqQtmXGZhvfDrKB.json",
     package = "rinspect"
  )

example_eval_log <- eval_log_read(file)

example_eval_log
}
