% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{inspect_log}
\alias{inspect_log}
\title{Write a task to an eval log}
\usage{
inspect_log(task, time_start = Sys.time(), dir = attr(res, "dir"))
}
\arguments{
\item{task}{An evaluation task created with \code{task_create()}.}

\item{time_start}{TODO: remove.}

\item{dir}{Directory where logs should be stored.}
}
\value{
The path to the directory which the log was written to. Pass this value
to \code{\link[=inspect_view]{inspect_view()}} to view the task logs.
}
\description{
This function translates a \link[=task_create]{task} to an evaluation log file
readable by the Inspect Log Viewer.

The usual entry point to this function is \code{\link[=inspect_view]{inspect_view()}}; use \code{\link[=inspect_log]{inspect_log()}}
to write a persistent log of your task eval. That said, evaluation logs
can't be read back into R as \code{task}s and later analyzed (or viewed in the
Inspect Log Viewer, for that matter); for that use, you likely want to
save the \code{task} using \code{\link[=save]{save()}} or \code{\link[=saveRDS]{saveRDS()}}.
}
