% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R, R/view.R
\name{task_log}
\alias{task_log}
\alias{inspect_view}
\alias{inspect_view.character}
\alias{inspect_view.task}
\title{The Inspect Log Viewer}
\usage{
task_log(task, time_start = Sys.time(), dir = attr(res, "dir"))

inspect_view(x, host = "127.0.0.1", port = 7576, ...)

\method{inspect_view}{character}(x, host = "127.0.0.1", port = 7576, ...)

\method{inspect_view}{task}(x, host = "127.0.0.1", port = 7576, ...)
}
\arguments{
\item{x}{Either a path to a directory containing a task eval log or a
task itself. If a task, the function will log the task to a temporary
directory and open that directory in the viewer.}

\item{host}{Host to serve on. Defaults to "127.0.0.1",}

\item{port}{Port to serve on. Defaults to 7576, one greater than the Python
implementation.}
}
\description{
The Inspect Log Viewer
}
