% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged.R
\name{logged}
\alias{logged}
\title{Log a function's evaluation}
\usage{
logged(fn, ..., fn_name = deparse(substitute(fn)))
}
\value{
A modified version of \code{fn()} with signature \code{...} and a named list return
value. That list's return value is:
\itemize{
\item \code{name}: A string representing the name of the function called.
\item \code{value}: The original return value of \code{fn()}.
\item \code{arguments}: Arguments passed to \code{fn()}, captured with \code{\link[rlang:list2]{rlang::dots_list()}}.
}
}
\description{
Inspect has a logging mechanism s.t. the names of functions called as well
as the arguments supplied are stored in the eval log.

\code{logged()} takes a function and maps it to a function with this sort of
"logged" return value. The function is applied to both internal functions
and user-supplied solvers/scorers/metrics.
}
\examples{
logged(mean)(x = 1:3)
}
\keyword{internal}
