% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{model_graded_qa}
\alias{model_graded_qa}
\title{Scorer: Evaluate a response using a model}
\usage{
model_graded_qa(
  template = NULL,
  instructions = NULL,
  grade_pattern = "(?i)GRADE\\\\s*:\\\\s*([CPI])(.*)$",
  partial_credit = FALSE,
  chat = ellmer::chat_claude()
)
}
\arguments{
\item{template}{Grading template to use--a \code{glue()} string which will take
substitutions \code{input}, \code{answer}, \code{criterion}, \code{instructions}.}

\item{instructions}{Grading instructions.}

\item{grade_pattern}{A regex pattern to extract the final grade from the
judge model's response.}

\item{partial_credit}{Whether to allow partial credit.}

\item{chat}{An ellmer chat used to grade the model output.}
}
\value{
A function that will grade model responses according to the given instructions.
}
\description{
Scorer: Evaluate a response using a model
}
\examples{
if (interactive()) {
  model_graded_qa()
}

}
\concept{scorers}
