% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinspect-package.R
\docType{package}
\name{rinspect-package}
\alias{rinspect}
\alias{rinspect-package}
\title{rinspect: Large Language Model Evaluation for R}
\description{
A port of 'Inspect', a widely adopted 'Python' framework for large language model evaluation. Specifically aimed at 'ellmer' users who want to measure the effectiveness of their LLM-based apps, the package includes facilities for prompt engineering, tool usage, multi-turn dialog, and model graded evaluations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/rinspect}
  \item \url{https://simonpcouch.github.io/rinspect/}
  \item Report bugs at \url{https://github.com/simonpcouch/rinspect/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
