% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer-model.R
\name{scorer_model}
\alias{scorer_model}
\alias{model_graded_qa}
\alias{model_graded_fact}
\title{Model-based scoring}
\usage{
model_graded_qa(
  template = NULL,
  instructions = NULL,
  grade_pattern = "(?i)GRADE\\\\s*:\\\\s*([CPI])(.*)$",
  partial_credit = FALSE,
  chat = NULL
)

model_graded_fact(
  template = NULL,
  instructions = NULL,
  grade_pattern = "(?i)GRADE\\\\s*:\\\\s*([CPI])(.*)$",
  partial_credit = FALSE,
  chat = NULL
)
}
\arguments{
\item{template}{Grading template to use--a \code{glue()} string which will take
substitutions \code{input}, \code{answer}, \code{criterion}, \code{instructions}.}

\item{instructions}{Grading instructions.}

\item{grade_pattern}{A regex pattern to extract the final grade from the
judge model's response.}

\item{partial_credit}{Whether to allow partial credit.}

\item{chat}{An ellmer chat used to grade the model output, e.g.
\code{\link[ellmer:chat_claude]{ellmer::chat_claude()}}.}
}
\value{
A function that will grade model responses according to the given instructions.
The returned function takes a \link[=task_solve]{solved task} and outputs a
2-element list, where the first element \code{scores} is a vector of scores with
length \code{nrow(task)} and the second is list of ellmer chats that led to the
scores, also with length \code{nrow(task)}.
The function that \code{model_graded_qa()}'s outputs can be passed directly to
\code{\link[=task_score]{task_score()}}.
}
\description{
Model-based scoring makes use of a model to score output from a solver.
\itemize{
\item \code{model_graded_qa()} scores how well a solver answers a question/answer task.
\item \code{model_graded_fact()} determines whether a solver includes a given fact
in its response.
}

The two scorers are quite similar in their implementation, but use a different
default \code{template} to evaluate correctness.
}
\examples{
# Quality assurance -----------------------------
if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  simple_addition <- tibble(
    input = c("What's 2+2?", "What's 2+3?"),
    target = c("4", "5")
  )

  tsk <- Task$new(
    dataset = simple_addition, 
    solver = generate(chat = chat_claude()), 
    scorer = model_graded_qa()
  )
  
  tsk$eval()
}

# Factual response -------------------------------
if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  r_history <- tibble(
    input = c(
      "Who created the R programming language?",
      "In what year was version 1.0 of R released?"
    ),
    target = c("Ross Ihaka and Robert Gentleman.", "2000.")
  )

  tsk <- Task$new(
    dataset = r_history, 
    solver = generate(chat = chat_claude()), 
    scorer = model_graded_fact()
  )
  
  tsk$eval()
}

}
\seealso{
\link{scorer_detect} for string detection-based scoring.
}
