% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_log}
\alias{task_log}
\title{Write a task to an eval log}
\usage{
task_log(task, time_start = Sys.time(), dir = attr(res, "dir"))
}
\arguments{
\item{time_start}{TODO: remove.}
}
\description{
Write a task to an eval log
}
