% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_score}
\alias{task_score}
\title{Scoring tasks}
\usage{
task_score(task, scorer)
}
\arguments{
\item{task}{An evaluation task created with \code{task_create()}.}

\item{scorer}{A function that evaluates how well the solver's return value
approximates the corresponding elements of \code{dataset$target}. See
\link[=scorer_model]{model-based scoring} for examples.}
}
\value{
A \code{task} object, which is a subclass of a tibble.
\code{task_score()} appends columns \code{score} and \code{scorer} to its input.
}
\description{
Evaluation \code{tasks} provide a flexible data structure for evaluating LLM-based
tools.
\enumerate{
\item \strong{Datasets} contain a set of labelled samples. Datasets are just a
tibble with columns \code{input} and \code{target}, where \code{input} is a prompt
and \code{target} is either literal value(s) or grading guidance. Situate datasets
inside of a task with \code{\link[=task_create]{task_create()}}.
\item \strong{Solvers} evaluate the \code{input} in the dataset and produce a final result.
The simplest solver is just an ellmer chat---evaluate a task with a solver
using \code{\link[=task_solve]{task_solve()}}.
\item \strong{Scorers} evaluate the final output of solvers. They may use text
comparisons (like \code{\link[=detect_match]{detect_match()}}), model grading (like \code{\link[=model_graded_qa]{model_graded_qa()}}),
or other custom schemes. Score solver results using \code{\link[=task_score]{task_score()}}.
}
}
\examples{
if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  simple_addition <- tibble(
    input = c("What's 2+2?", "What's 2+3?"),
    target = c("4", "5")
  )

  tsk <- task_create(dataset = simple_addition)
  tsk

  tsk <- task_solve(tsk, solver = chat_claude())
  tsk

  tsk <- task_score(tsk, scorer = model_graded_qa())
  tsk

  if (interactive()) {
    inspect_view(tsk)
  }
}

}
\seealso{
A typical evaluation with rinspect calls three functions in sequence:
\itemize{
\item Create an evaluation task with \code{\link[=task_create]{task_create()}}.
\item Generate solutions with \code{\link[=task_solve]{task_solve()}}.
\item Grade solutions with \code{\link[=task_score]{task_score()}}.
}

Then, see \code{\link[=inspect_log]{inspect_log()}} and \code{\link[=inspect_view]{inspect_view()}} for writing results to
file and interfacing with the Inspect Log Viewer.
}
