% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitals-package.R
\docType{package}
\name{vitals-package}
\alias{vitals}
\alias{vitals-package}
\title{vitals: Large Language Model Evaluation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A port of 'Inspect', a widely adopted 'Python' framework for large language model evaluation. Specifically aimed at 'ellmer' users who want to measure the effectiveness of their large language model-based products, the package supports prompt engineering, tool usage, multi-turn dialog, and model graded evaluations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverse/vitals}
  \item \url{https://vitals.tidyverse.org}
  \item Report bugs at \url{https://github.com/tidyverse/vitals/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Max Kuhn \email{max@posit.co} [contributor]
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID}) [contributor]
  \item Mine Cetinkaya-Rundel \email{mine@posit.co} (\href{https://orcid.org/0000-0001-6452-2420}{ORCID}) [contributor]
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
