% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{eval_log}
\alias{eval_log}
\alias{eval_log_read}
\alias{eval_log_write}
\alias{eval_log_dir}
\alias{eval_log_dir_set}
\title{Read and write evaluation logs}
\usage{
eval_log_read(x)

eval_log_write(x = eval_log_new(), dir = eval_log_dir())

eval_log_dir()

eval_log_dir_set(dir)
}
\arguments{
\item{x}{For \code{eval_log_read()}, a path to the log file to read. For
For \code{eval_log_write()}, an evaluation log object to write to disk.}

\item{dir}{Directory where logs should be stored.}
}
\description{
Utilities for reading and writing evaluation logs.

These utilities support the \code{INSPECT_LOG_DIR} environment variable,
which sets a default directory to write logs to.
}
\examples{
file <-
  system.file(
    "logs/2025-02-08T15-51-00-06-00_simple-arithmetic_o3cKtmsvqQtmXGZhvfDrKB.json",
     package = "rinspect"
  )

example_eval_log <- eval_log_read(file)

example_eval_log
}
