% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{generate}
\alias{generate}
\title{Solver: generate output}
\usage{
generate(chat = ellmer::chat_claude())
}
\arguments{
\item{chat}{An ellmer chat.}
}
\value{
A function that takes an \code{input} parameter and an optional \code{chat} parameter,
returning a named list containing the final \code{result} as well as the ellmer
\code{chat} object that generated it.
}
\description{
Solvers evaluate the \code{input} in a dataset and produce a result aimed at
replicating the \code{target}. \code{generate()} is the most basic solver in rinspect,
and just calls the model with the input and collects the output.
}
\examples{
generate()
}
\concept{solvers}
