% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{generate}
\alias{generate}
\title{Convert a chat to a solver function}
\usage{
generate(chat)
}
\arguments{
\item{chat}{An ellmer chat object, such as from \code{\link[ellmer:chat_claude]{ellmer::chat_claude()}}}
}
\value{
A solver function that can be used with Task
}
\description{
Convert a chat to a solver function
}
