% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{generate}
\alias{generate}
\title{Convert a chat to a solver function}
\usage{
generate(solver_chat = NULL)
}
\arguments{
\item{solver_chat}{An ellmer chat object, such as from \code{\link[ellmer:chat_anthropic]{ellmer::chat_anthropic()}}.}
}
\value{
The output of \code{generate()} is another function. That function takes in
a vector of \code{input}s, as well as a solver chat by the
name of \code{solver_chat} with the default supplied to \code{generate()} itself.

See the documentation for the \code{solver} argument in \link{Task} for more
information on the return type.
}
\description{
\code{generate()} is the simplest possible solver one might use with
vitals; it just passes its inputs to the supplied model and returns
its raw responses. The inputs are evaluated in parallel,
not in the sense of multiple R sessions, but in the sense of multiple,
asynchronous HTTP requests using \verb{$chat_parallel()}. \code{generate()}'s output
can be passed directory to the \code{solver} argument of \link{Task}'s \verb{$new()}
method.
}
