% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{inspect_log_dir}
\alias{inspect_log_dir}
\alias{inspect_log_dir_set}
\title{The log directory}
\usage{
inspect_log_dir()

inspect_log_dir_set(dir)
}
\value{
Both \code{inspect_log_dir()} and \code{inspect_log_dir_set()} return the current
value of the environment variable \code{INSPECT_LOG_DIR}. \code{inspect_log_dir_set()}
additionally sets it to a new value.
}
\description{
rinspect supports the \code{INSPECT_LOG_DIR} environment variable,
which sets a default directory to write logs to in \code{\link[=task_create]{task_create()}}
and \code{\link[=inspect_log]{inspect_log()}} and to view from in \code{\link[=inspect_view]{inspect_view()}}.
}
\examples{
inspect_log_dir()

dir <- tempdir()

inspect_log_dir_set(dir)

inspect_log_dir()
}
