% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{inspect_view}
\alias{inspect_view}
\title{Launch the Inspect Log Viewer}
\usage{
inspect_view(
  log_dir = system.file("logs", package = "rinspect"),
  host = "127.0.0.1",
  port = 7576
)
}
\arguments{
\item{log_dir}{Path to directory containing eval logs.
Defaults to package example logs.}

\item{host}{Host to serve on. Defaults to "127.0.0.1",}

\item{port}{Port to serve on. Defaults to 7576, one greater than the Python
implementation.}
}
\description{
Launch the Inspect Log Viewer
}
