% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{read_eval_log}
\alias{read_eval_log}
\title{Read and write evaluation logs}
\usage{
read_eval_log(file)
}
\description{
Read and write evaluation logs
}
\examples{
file <-
  system.file(
    "logs/2025-02-08T15-51-00-06-00_simple-arithmetic_o3cKtmsvqQtmXGZhvfDrKB.json",
     package = "rinspect"
  )

example_eval_log <- read_eval_log(file)

example_eval_log
}
