% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinspect-package.R
\docType{package}
\name{rinspect-package}
\alias{rinspect}
\alias{rinspect-package}
\title{rinspect: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
