% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer-model.R
\name{scorer_model}
\alias{scorer_model}
\alias{model_graded_qa}
\alias{model_graded_fact}
\title{Model-based scoring}
\usage{
model_graded_qa(
  template = NULL,
  instructions = NULL,
  grade_pattern = "(?i)GRADE\\\\s*:\\\\s*([CPI])(.*)$",
  partial_credit = FALSE,
  chat = NULL
)

model_graded_fact(
  template = NULL,
  instructions = NULL,
  grade_pattern = "(?i)GRADE\\\\s*:\\\\s*([CPI])(.*)$",
  partial_credit = FALSE,
  chat = NULL
)
}
\arguments{
\item{template}{Grading template to use--a \code{glue()} string which will take
substitutions \code{input}, \code{answer}, \code{criterion}, \code{instructions}.}

\item{instructions}{Grading instructions.}

\item{grade_pattern}{A regex pattern to extract the final grade from the
judge model's response.}

\item{partial_credit}{Whether to allow partial credit.}

\item{chat}{An ellmer chat used to grade the model output.}
}
\value{
A function that will grade model responses according to the given instructions.
This can be passed directly to \code{\link[=task_score]{task_score()}}.
}
\description{
Model-based scoring makes use of a model to score output from a solver.
\itemize{
\item \code{model_graded_qa()} scores how well a solver answers a question/answer task.
\item \code{model_graded_fact()} determines whether a solver includes a given fact
in its response.
}

The two scorers are quite similar in their implementation, but use a different
default \code{template} to evaluate correctness.
}
\examples{
# Quality assurance -----------------------------
if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  simple_addition <- tibble(
    input = c("What's 2+2?", "What's 2+3?"),
    target = c("4", "5")
  )

  tsk <- task_create(dataset = simple_addition)
  tsk

  tsk <- task_solve(tsk, solver = chat_claude())
  tsk

  tsk <- task_score(tsk, scorer = model_graded_qa())
  tsk

  if (interactive()) {
    inspect_view(tsk)
  }
}

# Factual response -------------------------------
if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  r_history <- tibble(
    input = c(
      "Who created the R programming language?",
      "In what year was version 1.0 of R released?"
    ),
    target = c("Ross Ihaka and Robert Gentleman.", "2000.")
  )

  tsk <- task_create(dataset = r_history)
  tsk

  tsk <- task_solve(tsk, solver = chat_claude())
  tsk

  tsk <- task_score(tsk, scorer = model_graded_fact())
  tsk

  if (interactive()) {
    inspect_view(tsk)
  }
}

}
\seealso{
\link{scorer_detect} for string detection-based scoring.
}
